Move tlm_adjoint/base_caches.py to tlm_adjoint/caches.py, tlm_adjoint/base_equations.py to tlm_adjoint/equations.py, and tlm_adjoint/numpy/numpy_equations.py and tlm_adjoint/numpy/equations.py. Update filenames in ACKNOWLEDGEMENTS.
