+++ tlm_adjoint/__init__.py
@@ -18,9 +18,9 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with tlm_adjoint.  If not, see <https://www.gnu.org/licenses/>.
 
-from .base_caches import *         # noqa: F401
-from .base_equations import *      # noqa: F401
+from .caches import *              # noqa: F401
 from .eigendecomposition import *  # noqa: F401
+from .equations import *           # noqa: F401
 from .functional import *          # noqa: F401
 from .hessian import *             # noqa: F401
 from .interface import *           # noqa: F401
