+++ tlm_adjoint/fenics/fenics_equations.py
@@ -26,8 +26,8 @@ from tlm_adjoint.interface import function_assign, function_comm, \
     real_function_value
 from .backend_code_generator_interface import assemble
 
-from tlm_adjoint.base_caches import Cache
-from tlm_adjoint.base_equations import Equation, EquationException, \
+from tlm_adjoint.caches import Cache
+from tlm_adjoint.equations import Equation, EquationException, \
     LinearEquation, Matrix, MatrixActionRHS, NullSolver, get_tangent_linear
 
 from .caches import form_dependencies, form_key
