+++ tlm_adjoint/firedrake/firedrake_equations.py
@@ -26,9 +26,9 @@ from tlm_adjoint.interface import function_assign, function_comm, \
     real_function_value
 from .backend_code_generator_interface import assemble, matrix_multiply
 
-from tlm_adjoint.base_caches import Cache
-from tlm_adjoint.base_equations import Equation, EquationException, \
-    NullSolver, get_tangent_linear
+from tlm_adjoint.caches import Cache
+from tlm_adjoint.equations import Equation, EquationException, NullSolver, \
+    get_tangent_linear
 
 from .caches import form_dependencies, form_key, parameters_key
 from .equations import EquationSolver, bind_form, derivative, unbind_form, \
