+++ tlm_adjoint/optimization.py
@@ -23,7 +23,7 @@ from .interface import function_axpy, function_copy, function_get_values, \
     function_linf_norm, function_local_size, function_new, \
     function_set_values, is_function
 
-from .base_caches import clear_caches
+from .caches import clear_caches
 from .manager import manager as _manager, set_manager
 
 import numpy as np
